Refactor Call settings menu and add a call chat dialog composable
