+++ dogfooding/src/main/kotlin/io/getstream/video/android/dogfooding/ui/call/CallChatDialog.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.video.android.dogfooding.ui.call
+
+import androidx.compose.runtime.Composable
+import io.getstream.video.android.core.Call
+
+@Composable
+internal fun CallChatDialog(
+    call: Call,
+    onDismissed: () -> Unit
+) {
+}
